+++ miaosha-common/miaosha-common-core/src/main/java/tech/tengshe789/miaosha/common/core/config/RestTemplateConfig.java
@@ -1,5 +1,6 @@
 package tech.tengshe789.miaosha.common.core.config;
 
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
@@ -13,6 +14,7 @@ import org.springframework.web.client.RestTemplate;
 @Configuration
 public class RestTemplateConfig {
 	@Bean
+	@LoadBalanced
 	public RestTemplate restTemplate ( ) {
 		return new RestTemplate();
 	}