+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/controller/LogController.java
@@ -1,0 +1,79 @@
+package tech.tengshe789.miaosha.sys.biz.controller;
+
+
+import com.baomidou.mybatisplus.core.toolkit.Wrappers;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import io.swagger.annotations.Api;
+import lombok.AllArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.*;
+import tech.tengshe789.miaosha.common.core.result.Result;
+import tech.tengshe789.miaosha.common.log.entity.SysLog;
+import tech.tengshe789.miaosha.common.security.annotation.Inner;
+import tech.tengshe789.miaosha.sys.biz.service.SysLogService;
+import tech.tengshe789.miaosha.sys.api.vo.PreLogVo;
+
+import javax.validation.Valid;
+import java.util.List;
+
+/**
+ * <p>
+ * 日志表 前端控制器
+ * </p>
+ *
+ * @author lengleng
+ */
+@RestController
+@AllArgsConstructor
+@RequestMapping("/log")
+@Api(value = "log", description = "日志管理模块")
+public class LogController {
+	private final SysLogService sysLogService;
+
+	/**
+	 * 简单分页查询
+	 *
+	 * @param page   分页对象
+	 * @param sysLog 系统日志
+	 * @return
+	 */
+	@GetMapping("/page")
+	public Result getLogPage(Page page, SysLog sysLog) {
+		return Result.success(sysLogService.page(page, Wrappers.query(sysLog)));
+	}
+
+	/**
+	 * 删除日志
+	 *
+	 * @param id ID
+	 * @return success/false
+	 */
+	@DeleteMapping("/{id}")
+	@PreAuthorize("@pms.hasPermission('sys_log_del')")
+	public Result removeById(@PathVariable Long id) {
+		return Result.success(sysLogService.removeById(id));
+	}
+
+	/**
+	 * 插入日志
+	 *
+	 * @param sysLog 日志实体
+	 * @return success/false
+	 */
+	@Inner
+	@PostMapping("/save")
+	public Result save(@Valid @RequestBody SysLog sysLog) {
+		return Result.success(sysLogService.save(sysLog));
+	}
+
+	/**
+	 * 批量插入前端异常日志
+	 *
+	 * @param preLogVoList 日志实体
+	 * @return success/false
+	 */
+	@PostMapping("/logs")
+	public Result saveBatchLogs(@RequestBody List<PreLogVo> preLogVoList) {
+		return Result.success(sysLogService.saveBatchLogs(preLogVoList));
+	}
+}