+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/controller/OauthClientDetailsController.java
@@ -1,0 +1,92 @@
+package tech.tengshe789.miaosha.sys.biz.controller;
+
+import com.baomidou.mybatisplus.core.toolkit.Wrappers;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import io.swagger.annotations.Api;
+import lombok.AllArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.*;
+import tech.tengshe789.miaosha.common.core.result.Result;
+import tech.tengshe789.miaosha.common.log.annotation.SysLog;
+import tech.tengshe789.miaosha.sys.api.entity.SysOauthClientDetails;
+import tech.tengshe789.miaosha.sys.biz.service.SysOauthClientDetailsService;
+
+import javax.validation.Valid;
+
+/**
+ * <p>
+ * 前端控制器
+ * </p>
+ *
+ * @author lengleng
+ */
+@RestController
+@AllArgsConstructor
+@RequestMapping("/client")
+@Api(value = "client", description = "客户端管理模块")
+public class OauthClientDetailsController {
+	private final SysOauthClientDetailsService sysOauthClientDetailsService;
+
+	/**
+	 * 通过ID查询
+	 *
+	 * @param id ID
+	 * @return SysOauthClientDetails
+	 */
+	@GetMapping("/{id}")
+	public Result getById(@PathVariable Integer id) {
+		return Result.success(sysOauthClientDetailsService.getById(id));
+	}
+
+
+	/**
+	 * 简单分页查询
+	 *
+	 * @param page                  分页对象
+	 * @param sysOauthClientDetails 系统终端
+	 * @return
+	 */
+	@GetMapping("/page")
+	public Result getOauthClientDetailsPage(Page page, SysOauthClientDetails sysOauthClientDetails) {
+		return new Result<>(sysOauthClientDetailsService.page(page, Wrappers.query(sysOauthClientDetails)));
+	}
+
+	/**
+	 * 添加
+	 *
+	 * @param sysOauthClientDetails 实体
+	 * @return success/false
+	 */
+	@SysLog("添加终端")
+	@PostMapping
+	@PreAuthorize("@pms.hasPermission('sys_client_add')")
+	public Result add(@Valid @RequestBody SysOauthClientDetails sysOauthClientDetails) {
+		return Result.success(sysOauthClientDetailsService.save(sysOauthClientDetails));
+	}
+
+	/**
+	 * 删除
+	 *
+	 * @param id ID
+	 * @return success/false
+	 */
+	@SysLog("删除终端")
+	@DeleteMapping("/{id}")
+	@PreAuthorize("@pms.hasPermission('sys_client_del')")
+	public Result removeById(@PathVariable String id) {
+		return Result.success(sysOauthClientDetailsService.removeClientDetailsById(id));
+	}
+
+	/**
+	 * 编辑
+	 *
+	 * @param sysOauthClientDetails 实体
+	 * @return success/false
+	 */
+	@SysLog("编辑终端")
+	@PutMapping
+	@PreAuthorize("@pms.hasPermission('sys_client_edit')")
+	public Result update(@Valid @RequestBody SysOauthClientDetails sysOauthClientDetails) {
+		return Result.success(sysOauthClientDetailsService.updateClientDetailsById(sysOauthClientDetails));
+	}
+}