+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/controller/RoleController.java
@@ -1,0 +1,97 @@
+package tech.tengshe789.miaosha.sys.biz.controller;
+
+import com.baomidou.mybatisplus.core.toolkit.Wrappers;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import io.swagger.annotations.Api;
+import lombok.AllArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.*;
+import tech.tengshe789.miaosha.common.core.result.Result;
+import tech.tengshe789.miaosha.common.log.annotation.SysLog;
+import tech.tengshe789.miaosha.sys.api.entity.SysRole;
+import tech.tengshe789.miaosha.sys.biz.service.SysRoleService;
+
+import javax.validation.Valid;
+
+/**
+ * @author lengleng
+ * @date 2017/11/5
+ */
+@RestController
+@AllArgsConstructor
+@RequestMapping("/role")
+@Api(value = "role", description = "角色管理模块")
+public class RoleController {
+	private final SysRoleService sysRoleService;
+
+	/**
+	 * 通过ID查询角色信息
+	 *
+	 * @param id ID
+	 * @return 角色信息
+	 */
+	@GetMapping("/{id}")
+	public Result getById(@PathVariable Integer id) {
+		return Result.success(sysRoleService.getById(id));
+	}
+
+	/**
+	 * 添加角色
+	 *
+	 * @param sysRole 角色信息
+	 * @return success、false
+	 */
+	@SysLog("添加角色")
+	@PostMapping
+	@PreAuthorize("@pms.hasPermission('sys_role_add')")
+	public Result save(@Valid @RequestBody SysRole sysRole) {
+		return Result.success(sysRoleService.save(sysRole));
+	}
+
+	/**
+	 * 修改角色
+	 *
+	 * @param sysRole 角色信息
+	 * @return success/false
+	 */
+	@SysLog("修改角色")
+	@PutMapping
+	@PreAuthorize("@pms.hasPermission('sys_role_edit')")
+	public Result update(@Valid @RequestBody SysRole sysRole) {
+		return Result.success(sysRoleService.updateById(sysRole));
+	}
+
+	/**
+	 * 删除角色
+	 *
+	 * @param id
+	 * @return
+	 */
+	@SysLog("删除角色")
+	@DeleteMapping("/{id}")
+	@PreAuthorize("@pms.hasPermission('sys_role_del')")
+	public Result removeById(@PathVariable Integer id) {
+		return Result.success(sysRoleService.removeRoleById(id));
+	}
+
+	/**
+	 * 获取角色列表
+	 *
+	 * @return 角色列表
+	 */
+	@GetMapping("/list")
+	public Result listRoles() {
+		return Result.success(sysRoleService.list(Wrappers.emptyWrapper()));
+	}
+
+	/**
+	 * 分页查询角色信息
+	 *
+	 * @param page 分页对象
+	 * @return 分页对象
+	 */
+	@GetMapping("/page")
+	public Result getRolePage(Page page) {
+		return Result.success(sysRoleService.page(page, Wrappers.emptyWrapper()));
+	}
+}