+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/mapper/SysLogMapper.java
@@ -1,0 +1,14 @@
+package tech.tengshe789.miaosha.sys.biz.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import tech.tengshe789.miaosha.common.log.entity.SysLog;
+
+/**
+ * <p>
+ * 日志表 Mapper 接口
+ * </p>
+ *
+ * @author lengleng
+ */
+public interface SysLogMapper extends BaseMapper<SysLog> {
+}