+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/mapper/SysOauthClientDetailsMapper.java
@@ -1,0 +1,15 @@
+package tech.tengshe789.miaosha.sys.biz.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import tech.tengshe789.miaosha.sys.api.entity.SysOauthClientDetails;
+
+/**
+ * <p>
+ * Mapper 接口
+ * </p>
+ *
+ * @author lengleng
+ */
+public interface SysOauthClientDetailsMapper extends BaseMapper<SysOauthClientDetails> {
+
+}