+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/mapper/SysRouteConfMapper.java
@@ -1,0 +1,14 @@
+package tech.tengshe789.miaosha.sys.biz.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import tech.tengshe789.miaosha.sys.api.entity.SysRouteConf;
+
+/**
+ * 路由
+ *
+ * @author lengleng
+ * @date 2018-11-06 10:17:18
+ */
+public interface SysRouteConfMapper extends BaseMapper<SysRouteConf> {
+
+}