+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/mapper/SysSocialDetailsMapper.java
@@ -1,0 +1,11 @@
+package tech.tengshe789.miaosha.sys.biz.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import tech.tengshe789.miaosha.sys.api.entity.SysSocialDetails;
+
+/**
+ * 系统社交登录账号表
+ */
+public interface SysSocialDetailsMapper extends BaseMapper<SysSocialDetails> {
+
+}