+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/service/SysOauthClientDetailsService.java
@@ -1,0 +1,29 @@
+package tech.tengshe789.miaosha.sys.biz.service;
+
+import com.baomidou.mybatisplus.extension.service.IService;
+import tech.tengshe789.miaosha.sys.api.entity.SysOauthClientDetails;
+
+/**
+ * <p>
+ * 服务类
+ * </p>
+ *
+ * @author lengleng
+ */
+public interface SysOauthClientDetailsService extends IService<SysOauthClientDetails> {
+	/**
+	 * 通过ID删除客户端
+	 *
+	 * @param id
+	 * @return
+	 */
+	Boolean removeClientDetailsById(String id);
+
+	/**
+	 * 根据客户端信息
+	 *
+	 * @param sysOauthClientDetails
+	 * @return
+	 */
+	Boolean updateClientDetailsById(SysOauthClientDetails sysOauthClientDetails);
+}