+++ miaosha-sys/miaosha-sys-biz/src/main/java/tech/tengshe789/miaosha/sys/biz/service/SysUserService.java
@@ -1,0 +1,73 @@
+package tech.tengshe789.miaosha.sys.biz.service;
+
+import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.baomidou.mybatisplus.extension.service.IService;
+import tech.tengshe789.miaosha.common.core.result.Result;
+import tech.tengshe789.miaosha.sys.api.dto.UserDTO;
+import tech.tengshe789.miaosha.sys.api.dto.UserInfo;
+import tech.tengshe789.miaosha.sys.api.entity.SysUser;
+import tech.tengshe789.miaosha.sys.api.vo.UserVO;
+
+/**
+ * @author lengleng
+ */
+public interface SysUserService extends IService<SysUser> {
+
+	/**
+	 * 删除用户
+	 *
+	 * @param sysUser 用户
+	 * @return boolean
+	 */
+	Boolean deleteUserById(SysUser sysUser);
+
+	/**
+	 * 更新当前用户基本信息
+	 *
+	 * @param userDto 用户信息
+	 * @return Boolean
+	 */
+	Result<Boolean> updateUserInfo(UserDTO userDto);
+
+	/**
+	 * 更新指定用户信息
+	 *
+	 * @param userDto 用户信息
+	 * @return
+	 */
+	Boolean updateUser(UserDTO userDto);
+
+	/**
+	 * 通过ID查询用户信息
+	 *
+	 * @param id 用户ID
+	 * @return 用户信息
+	 */
+	UserVO selectUserVoById(Long id);
+
+	/**
+	 * 保存用户信息
+	 *
+	 * @param userDto DTO 对象
+	 * @return success/fail
+	 */
+	Boolean saveUser(UserDTO userDto);
+
+	/**
+	 * 查询用户信息
+	 *
+	 * @param sysUser 用户
+	 * @return userInfo
+	 */
+	UserInfo findUserInfo(SysUser sysUser);
+
+	/**
+	 * 分页查询用户信息（含有角色信息）
+	 *
+	 * @param page    分页对象
+	 * @param userDTO 参数列表
+	 * @return
+	 */
+	IPage getUsersWithRolePage(Page page, UserDTO userDTO);
+}